# containers/eas/openapi.yaml
...
openapi: 3.0.3
...
info:
  # independent of eas/VERSION
  version: 0.2.0
  title: Entity Attribute Service
  termsOfService: https://www.virtru.com/terms-of-service/
  contact:
    email: support@virtru.com
  license:
    name: MIT
    url: https://github.com/virtru/etheria/blob/master/LICENSE
paths:
...
  /v1/attribute:
    post:
      tags:
        - attribute
      deprecated: True
      summary: Create an attribute using an attribute URI. The other values will be set to defaults. see eas/README.md
      operationId: src.web.attribute.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AttributeURI'
            examples:
              https:
                value: ["https://eas.local/attr/Classification/value/Unclassified"]
              many:
                value: ["https://eas.local/attr/Kingdom/value/Animalia","https://eas.local/attr/Phylum/value/Chordata","https://eas.local/attr/Class/value/Mammalia","https://eas.local/attr/Order/value/Primates","https://eas.local/attr/Family/value/Hominidae","https://eas.local/attr/Genus/value/Homo","https://eas.local/attr/Species/value/Homo-sapiens"]
      responses:
        "200":
          description: OK
    put:
      tags:
        - attribute
      deprecated: True
      summary: Update the properties of an attribute value.  see eas/README.md
      operationId: src.web.attribute.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Attribute"
            examples:
              minimum:
                value:
                  attribute: "https://eas.local/attr/Kingdom/value/Animalia"
                  kasUrl: "http://kas.local"
                  pubKey: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA2Q9axUqaxEfhOO2+0Xw+\nswa5Rb2RV0xeTX3GC9DeORv9Ip49oNy+RXvaMsdNKspPWYZZEswrz2+ftwcQOSU+\nefRCbGIwbSl8QBfKV9nGLlVmpDydcAIajc7YvWjQnDTEpHcJdo9y7/oogG7YcEmq\nS3NtVJXCmbc4DyrZpc2BmZD4y9417fSiNjTTYY3Fc19lQz07hxDQLgMT21N4N0Fz\nmD6EkiEpG5sdpDT/NIuGjFnJEPfqIs6TnPaX2y1OZ2/JzC+mldJFZuEqJZ/6qq/e\nYlp04nWrSnXhPpTuxNZ5J0GcPbpcFgdT8173qmm5m5jAjiFCr735lH7USl15H2fW\nTwIDAQAB\n-----END PUBLIC KEY-----\n"
              maximum:
                value:
                  attribute: "https://eas.local/attr/Phylum/value/Chordata"
                  kasUrl: "http://kas.local"
                  pubKey: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA2Q9axUqaxEfhOO2+0Xw+\nswa5Rb2RV0xeTX3GC9DeORv9Ip49oNy+RXvaMsdNKspPWYZZEswrz2+ftwcQOSU+\nefRCbGIwbSl8QBfKV9nGLlVmpDydcAIajc7YvWjQnDTEpHcJdo9y7/oogG7YcEmq\nS3NtVJXCmbc4DyrZpc2BmZD4y9417fSiNjTTYY3Fc19lQz07hxDQLgMT21N4N0Fz\nmD6EkiEpG5sdpDT/NIuGjFnJEPfqIs6TnPaX2y1OZ2/JzC+mldJFZuEqJZ/6qq/e\nYlp04nWrSnXhPpTuxNZ5J0GcPbpcFgdT8173qmm5m5jAjiFCr735lH7USl15H2fW\nTwIDAQAB\n-----END PUBLIC KEY-----\n"
                  isDefault: false
                  displayName: "Phylum::Chordata"
      responses:
        "200":
          $ref: "#/components/responses/AttributeArray"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/attr:
    parameters:
      - name: namespace
        in: query
        description: Authority namespace. overrides default
        schema:
          $ref: "#/components/schemas/AuthorityNamespace"
    head:
      tags:
        - attribute name
      summary: (Header only) Return all attribute name objects. Attribute names are comprised of a classification, a set of possible values, and a rule.
      operationId: src.web.attribute_name.head
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags:
        - attribute name
      summary: Return all attribute name objects. Attribute names are comprised of a classification, a set of possible values, and a rule.
      operationId: src.web.attribute_name.find
      responses:
        "200":
          $ref: "#/components/responses/AttributeNameObjectArray"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags:
        - attribute name
      summary: Create attribute name(s), which exists in an authority namespace and is comprised a classification, a set of possible values, and a rule.
      description: 'If attribute name cannot be created (for example if body is malformed), [uri: null] will be returned.'
      operationId: src.web.attribute_name.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AttributeNameObject"
            example:
              - authorityNamespace: "https://eas.local"
                name: "Kingdom"
                rule: allOf
              - authorityNamespace: "https://eas.local"
                name: "Phylum"
                rule: allOf
      responses:
        "201":
          $ref: "#/components/responses/AttributeNameObjectArray"
        "400":
          $ref: "#/components/responses/BadRequest"
    patch:
      tags:
        - attribute name
      summary: Patch attribute name object(s) - merges the new data into existing objects.
      description: 'If an attribute name is not found or body is malformed, [uri: null] will be returned.'
      operationId: src.web.attribute_name.patch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AttributeNameObject"
            example:
              - authorityNamespace: "https://eas.local"
                name: "Kingdom"
                rule: allOf
              - authorityNamespace: "https://eas.local"
                name: "Phylum"
                rule: allOf
      responses:
        "200":
          $ref: "#/components/responses/AttributeNameObjectArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/attrName:
    post:
      # Note this endpoint breaks the design pattern where post methods create. It's here because KAS needs a robust endpoint for attribute name discovery.
      tags:
        - attribute name
      summary: Return multiple attribute name objects, which exists in an authority namespace and is comprised a classification, a set of possible values, and a rule.
      operationId: src.web.attribute_name.get_many
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: path
            examples:
              none:
                value: []
              one:
                value: [ "https://eas.local/attr/Kingdom" ]
              many:
                value: [ "https://eas.local/attr/Kingdom","https://eas.local/attr/Phylum","https://eas.local/attr/Class","https://eas.local/attr/Order","https://eas.local/attr/Family","https://eas.local/attr/Genus","https://eas.local/attr/Species" ]
      responses:
        "200":
          $ref: "#/components/responses/AttributeNameObjectArray"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/attr/{name}:
    parameters:
      - name: name
        in: path
        description: Name of the attribute
        required: true
        schema:
          type: string
          example: "IntellectualProperty"
      - name: namespace
        in: query
        schema:
          $ref: "#/components/schemas/AuthorityNamespace"
    get:
      tags:
        - attribute name
      summary: Return details for an attribute name, which exists in an authority namespace and is comprised a classification, a set of possible values, and a rule.
      operationId: src.web.attribute_name.get
      responses:
        "200":
          $ref: "#/components/responses/AttributeNameObject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - attribute name
      summary: Update attribute name object. Idempotent
      operationId: src.web.attribute_name.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttributeNameObject"
            examples:
              minimum:
                value:
                  authorityNamespace: "https://eas.local"
                  name: "IntellectualProperty"
                  rule: "anyOf"
              maximum:
                value:
                  authorityNamespace: "https://eas.local"
                  name: "IntellectualProperty"
                  rule: "anyOf"
                  order: ["TradeSecret", "Proprietary", "BusinessSensitive", "Open"]
      responses:
        "200":
          $ref: "#/components/responses/AttributeNameObject"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/authorityNamespace:
    get:
      tags:
        - authority namespace
      summary: Return the authority namespaces supported by this EAS.
      operationId: src.web.authority_namespace.get
      parameters:
        - name: isDefault
          description: true to get the default namespace
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/AuthorityNamespaceArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - authority namespace
      summary: Create authority namespace
      operationId: src.web.authority_namespace.create
      requestBody:
        description: Authority namespace added and default state
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorityNamespaceObject"
      responses:
        "201":
          $ref: "#/components/responses/AuthorityNamespace"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  responses:
...
    AttributeArray:
      description: Respond with the attributes and its properties
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Attribute"
    AttributeNameObject:
      description: Respond with the attribute name
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AttributeNameObject"
    AttributeNameObjectArray:
      description: Respond with an array of attribute names
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/AttributeNameObject"
...
    AuthorityNamespaceArray:
      description: Respond with the authority namespace supported by this EAS.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/AuthorityNamespace"
...
    BadRequest:
      description: 400 Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServerError:
      description: 500 Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: 404 Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    AuthorityNamespace:
      description: authority namespace
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthorityNamespaceObject"
...
  schemas:
    Attribute:
      type: object
      required:
        - attribute
        - kasUrl
        - pubKey
      properties:
        attribute:
          $ref: "#/components/schemas/AttributeURI"
        displayName:
          type: string
          example: "Top Secret"
        pubKey:
          $ref: "#/components/schemas/PublicKey"
        kasUrl:
          description: Base URL of a KAS that can make access control decisions for this attribute.
          type: string
          format: path
          example: "http://kas.local"
        isDefault:
          description: If true this flag identifies the attribute as the default attribute. If missing (preferred) or false then the attribute is not the default attribute. (Optional)
          type: boolean
          example: false
        schemaVersion:
          description: Version number of the TDF3 Attribute Object schema (optional).
          type: string
          example: "1.1.0"
        state:
          type: string
          enum: [active, inactive]
    AttributeURI:
      description: The unique and immutable resource name for the attribute represented as a case-sensitive URI of the format https://[auth namespace]/attr/[attr name]/value/[attr value]
      type: string
      format: path
      example: "https://eas.local/attr/ClassificationUS/value/Unclassified"
...
    AttributeNameObject:
      description: An authority namespace, name and rule.
      type: object
      required:
        - authorityNamespace
        - name
        - rule
      properties:
        authorityNamespace:
          type: string
          example: "https://eas.local"
        name:
          description: Name of the attribute, e.g. 'color' in 'https://eas.local/attr/color/value/green'
          type: string
          example: "IntellectualProperty"
        rule:
          type: string
          enum: [anyOf, allOf, hierarchy]
        order:
          type: array
          items:
            type: string
          example: ["TradeSecret", "Proprietary", "BusinessSensitive", "Open"]
        state:
          type: string
          enum: [active, inactive]
    AuthorityNamespace:
      description: A canonical domain responsible for managing attribute names (and thus their attribute values)
      type: string
      format: path
      example: "https://eas.local"
    AuthorityNamespaceObject:
      description: s
      type: object
      properties:
        namespace:
          $ref: "#/components/schemas/AuthorityNamespace"
        isDefault:
          type: boolean
...
    Problem:
      description: Standard problem details message used for errors. From https://tools.ietf.org/html/draft-ietf-appsawg-http-problem-00 .
      type: object
      properties:
        title:
          type: string
          description: Human-readable summary of problem
          example: "Exception name"
        detail:
          type: string
          description: An human readable explanation specific to this
            occurrence of the problem.
          example: "An human readable explanation specific to this occurrence of the problem."
        status:
          type: integer
          default: 500
          example: 000
        type:
          type: string
          description: Link to web page with further information on the problem, if available.
          example: "https://eas.example.com/problem_information"
          default: about:blank
    PublicKey:
      type: string
      example: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA2Q9axUqaxEfhOO2+0Xw+\nswa5Rb2RV0xeTX3GC9DeORv9Ip49oNy+RXvaMsdNKspPWYZZEswrz2+ftwcQOSU+\nefRCbGIwbSl8QBfKV9nGLlVmpDydcAIajc7YvWjQnDTEpHcJdo9y7/oogG7YcEmq\nS3NtVJXCmbc4DyrZpc2BmZD4y9417fSiNjTTYY3Fc19lQz07hxDQLgMT21N4N0Fz\nmD6EkiEpG5sdpDT/NIuGjFnJEPfqIs6TnPaX2y1OZ2/JzC+mldJFZuEqJZ/6qq/e\nYlp04nWrSnXhPpTuxNZ5J0GcPbpcFgdT8173qmm5m5jAjiFCr735lH7USl15H2fW\nTwIDAQAB\n-----END PUBLIC KEY-----\n"
...